Add Configuration for exporting GeoJSON and TopoJSON . Including Deletion of TopoJSON

+++ gushim/program.py
@@ -23,8 +23,9 @@ END_ATTRIBUTE_FILE = 'attributes01.csv'
 YESHUV_MASK_FILE = r'support_data/Yeshuvim2015.csv'
 MIN_POPULATION = 2000
 SAVE_GUSHIM_SHAPEFILE = True  #save the gushim to shapefile
-CONVERT_GEOJSON2TOPOJSON = True  #Save also to TopoJSON
-
+EXPORT_TO_GEOJSON = False
+EXPORT_TO_TOPOJSON = True  #Save also to TopoJSON
+DELETE_GEOJSON = True
 
 def setup_logging(default_path='logging.yaml', default_level=logging.INFO, env_key='LOG_CFG'):
     """
@@ -152,12 +153,22 @@ def main():
             file_name_string = "".join(x for x in local if x.isalnum()).encode('utf-8')  #remove unsafe characters
             try:
                 export_file = os.path.join(base_folder, export_folder, 'local_{0}.geojson'.format(file_name_string))
-                with open(export_file, 'w') as f:
-                    f.write(df_local.to_json())
-                if CONVERT_GEOJSON2TOPOJSON:
+                local_geojson = df_local.to_json()
+                if EXPORT_TO_GEOJSON:
+                    with open(export_file, 'w') as f:
+                        f.write(local_geojson)
+                if EXPORT_TO_TOPOJSON:
+                    if not os.path.exists(export_file) or not os.path.isdir(export_file):
+                        with open(export_file, 'w') as f:
+                            f.write(local_geojson)
                     filename, _ = os.path.splitext(export_file)
                     topojson_file = filename + '.topojson'
                     geo_utils.geoson_to_topojson(export_file, topojson_file)
+                if DELETE_GEOJSON:
+                    try:
+                        os.remove(export_file)
+                    except OSError, e:
+                        print ("Error: {} - {}.".format(e.export_file, e.strerror))
             except:
                 logger.warning('failed to save geojson for: {0}'.format(local))
     logger.debug('Finished to saved files to GeoJSON')
@@ -179,7 +190,15 @@ def main():
     # geojson_file = r'C:\Users\sephi\github\opentaba_gushim\opentaba_gushim_prj\gushim\workspace\abugosh.geoson'
     #
     #Generate Topojson (reuse)
-    #Push to GIT
+
+    # Log Statistics
+    logger.info('Number of Localities to be exported is {}'.format())
+    logger.info('Number of Gushim / Polygons that are in to be s {}'.format())
+    logger.info('Number of Localities to be exported is {}'.format())
+    logger.info('Number of Localities to be exported is {}'.format())
+
+    # TODO Push to GIT
+
     logger.info("Program completed")
